Overlay in gameplay works
need to add a check for completed text to speech now
