--- StoryQuest/app/Continue/page.tsx
@@ -1,48 +1,0 @@
-"use client";
-
-import "../CompletionPage/CompletionPageStyling.css";
-import React, {useState} from "react";
-import Image from "next/image";
-import "@/CreateRoom/CreateRoomButtonStyles.css";
-import useSound from "use-sound";
-import CompletionPage from "../CompletionPage/page.tsx"; // Import CompletionPage
-
-
-
-export default function Home() {
-    const popClick = '/sounds/pop-click.mp3';
-    const [play] = useSound(popClick); // use sound hook, play sound
-    const [storyCompleted, setStoryCompleted] = useState(false);
-
-    return (
-/* this pops up after the completed text to speech is done after about 1 sec delay*/
-        <div className="page-container">
-                    {!storyCompleted && (
-                        <>
-                        <div className="content-container align-container">
-                        <h1 className="text-color">Are you completed with the story?</h1>
-                            <button
-                                className="button setup-room-button"
-                                onClick={() => setStoryCompleted(true)}
-                            >
-                                <Image
-                                    src="/arrow-icon.svg"
-                                    alt="Continue icon"
-                                    width={30}
-                                    height={30}
-                                    className="icon-spacing"
-                                />
-                                <span>Continue</span>
-                            </button>
-                            </div>
-                        </>
-                    )}
-                    {storyCompleted && (
-                        <div className="overlay">
-                            <CompletionPage />
-                        </div>
-                    )}
-        </div>
-
-    );
-}
